feat: refactor settings context
