+++ ios/SmartVideo-Info.plist
@@ -9,78 +9,10 @@
 		<dict>
 			<key>Environment</key>
 			<string>live</string>
-			<key>Engine</key>
-			<string>generic</string>
-			<key>SmartVideo URL</key>
-			<string>videome.videoengager.com</string>
 		</dict>
 		<dict>
-			<key>SmartVideo URL</key>
-			<string>staging.videoengager.com</string>
 			<key>Environment</key>
 			<string>staging</string>
-			<key>agents</key>
-			<array>
-				<dict>
-					<key>Tenant ID</key>
-					<string>oIiTR2XQIkb7p0ub</string>
-					<key>Name</key>
-					<string>mobiledev</string>
-				</dict>
-			</array>
-			<key>Environment URL</key>
-			<string>https://api.mypurecloud.de</string>
-			<key>Queue Name</key>
-			<string>Support</string>
-			<key>Organization ID</key>
-			<string>639292ca-14a2-400b-8670-1f545d8aa860</string>
-			<key>Engine</key>
-			<string>genesys</string>
-			<key>Deployment ID</key>
-			<string>1b4b1124-b51c-4c38-899f-3a90066c76cf</string>
-		</dict>
-		<dict>
-			<key>SmartVideo URL</key>
-			<string>videome.videoengager.com</string>
-			<key>Environment</key>
-			<string>live</string>
-			<key>agents</key>
-			<array>
-				<dict>
-					<key>Customer Label</key>
-					<string>Dummy Display Name</string>
-					<key>Tenant ID</key>
-					<string>0FphTk091nt7G1W7</string>
-					<key>Name</key>
-					<string>mobiledev</string>
-				</dict>
-			</array>
-			<key>Environment URL</key>
-			<string>https://api.mypurecloud.com</string>
-			<key>Queue Name</key>
-			<string>Support</string>
-			<key>Organization ID</key>
-			<string>c4b553c3-ee42-4846-aeb1-f0da3d85058e</string>
-			<key>Engine</key>
-			<string>genesys</string>
-			<key>Deployment ID</key>
-			<string>973f8326-c601-40c6-82ce-b87e6dafef1c</string>
-		</dict>
-		<dict>
-			<key>SmartVideo URL</key>
-			<string>videome.leadsecure.com</string>
-			<key>Environment</key>
-			<string>live</string>
-			<key>Engine</key>
-			<string>genesysEngage</string>
-			<key>Service Name</key>
-			<string>customer-support</string>
-			<key>Server Base URL</key>
-			<string>https://briant-gme.live.genesys.com</string>
-			<key>Agent URL</key>
-			<string>engagepoc</string>
-			<key>Authorization Value</key>
-			<string></string>
 		</dict>
 	</array>
 </dict>
